+++ packages/cli/src/tasks/server.js
@@ -15,6 +15,23 @@ const startServer = async config => {
     server.openBrowser ? args.push('--open-browser') : args.push('--no-open-browser')
   }
 
+  if (server.auth?.users?.length) {
+    const { users, rules } = server.auth
+    users.forEach(user => {
+      const keys = Object.keys(user)
+      if (keys.length == 1) {
+        args.push('--user', `${keys[0]}:${user[keys[0]]}`)
+      }
+    })
+    if (rules?.length) {
+      rules.forEach(rule => {
+        const keys = Object.keys(rule)
+        if (keys.length == 1) {
+          args.push('--rule', `${keys[0]}:${rule[keys[0]]}`)
+        }
+      })
+    }
+  }
   await runCli(args)
   return 'exit'
 }