+++ packages/server/src/index.js
@@ -5,7 +5,6 @@ const cors = require('cors');
 const bodyParser = require('body-parser');
 const open = require('open')
 
-const logger = require('@home-gallery/logger')
 const { callbackify } = require('@home-gallery/common')
 
 const { loggerMiddleware } = require('./logger-middleware')
@@ -13,6 +12,8 @@ const { EventBus } = require('./eventbus');
 const databaseApi = require('./api/database');
 const eventsApi = require('./api/events');
 const webapp = require('./webapp');
+const { createBasicAuthMiddleware } = require('./auth')
+const { isIndex, skipIf } = require('./utils')
 
 const log = require('@home-gallery/logger')('server')
 const openCb = callbackify(open)
@@ -44,28 +45,40 @@ function createServer(key, cert, app) {
 
 const serverUrl = (port, key, cert) =>`${key && cert ? 'https' : 'http'}://localhost:${port}`
 
+const getAuthMiddleware = options => {
+  if (options.users?.length) {
+    return createBasicAuthMiddleware(options.users, options.ipWhitelistRules || defaultIpWhitelistRules)
+  }
+  return (req, _, next) => next()
+}
+
 function startServer(options, cb) {
   const {host, port, storageDir, databaseFilename, eventsFilename, webappDir, key, cert, openBrowser} = options
   const app = express();
-  app.disable('x-powered-by');
+  app.disable('x-powered-by')
+  app.enable('trust proxy')
 
   app.use(loggerMiddleware())
   app.use(cors());
   app.use(compression({ filter: shouldCompress }))
-  app.use('/files', express.static(storageDir, {index: false, maxAge: '2d', immutable: true}));
 
+  app.use(skipIf(express.static(webappDir), isIndex))
+
+  app.use(getAuthMiddleware(options))
+
+  app.use('/files', express.static(storageDir, {index: false, maxAge: '2d', immutable: true}));
   app.use(bodyParser.json({limit: '1mb'}))
 
-  const { read, push, stream, getEvents } = eventsApi(eventbus, eventsFilename);
-  const { read: dbRead, init: dbInit, getFirstEntries } = databaseApi(eventbus);
-  app.get('/api/database.json', dbRead);
-  app.get('/api/events.json', read);
+  const { read: readEvents, push: pushEvent, stream, getEvents } = eventsApi(eventbus, eventsFilename);
+  const { read: readDatabase, init: initDatabase, getFirstEntries } = databaseApi(eventbus, databaseFilename, getEvents);
+  app.get('/api/events.json', readEvents);
   app.get('/api/events/stream', stream);
-  app.post('/api/events', push);
+  app.post('/api/events', pushEvent);
+  app.get('/api/database.json', readDatabase);
 
   // deprecated
-  app.get('/api/database', dbRead);
-  app.get('/api/events', read);
+  app.get('/api/database', readDatabase);
+  app.get('/api/events', readEvents);
 
   app.use(webapp(webappDir, getFirstEntries, 50));
 
@@ -80,7 +93,7 @@ function startServer(options, cb) {
     .on('listening', () => {
       const url = serverUrl(port, key, cert)
       log.info(`Open Home Gallery on ${url}`);
-      dbInit(databaseFilename, getEvents);
+      initDatabase();
       if (openBrowser) {
         log.debug(`Open browser with url ${url}`)
         return openCb(url, () => cb(null, app))